Consolidate many dataclasses into model.py.
