+++ prymer/__init__.py
@@ -1,0 +1,7 @@
+from prymer.model import MinOptMax
+from prymer.model import Oligo
+from prymer.model import PrimerPair
+from prymer.model import Span
+from prymer.model import Strand
+
+__all__ = ["Strand", "Span", "Oligo", "PrimerPair", "MinOptMax"]
