+++ prymer/api/__init__.py
@@ -1,10 +1,4 @@
-from prymer.api.minoptmax import MinOptMax
-from prymer.api.oligo import Oligo
 from prymer.api.picking import build_primer_pairs
-from prymer.api.primer_pair import PrimerPair
-from prymer.api.span import BedLikeCoords
-from prymer.api.span import Span
-from prymer.api.span import Strand
 from prymer.api.variant_lookup import FileBasedVariantLookup
 from prymer.api.variant_lookup import SimpleVariant
 from prymer.api.variant_lookup import VariantLookup
@@ -14,13 +8,7 @@ from prymer.api.variant_lookup import cached
 from prymer.api.variant_lookup import disk_based
 
 __all__ = [
-    "MinOptMax",
     "build_primer_pairs",
-    "Oligo",
-    "PrimerPair",
-    "Span",
-    "Strand",
-    "BedLikeCoords",
     "VariantType",
     "SimpleVariant",
     "VariantLookup",
