+++ tests/api/test_picking.py
@@ -7,12 +7,12 @@ import pytest
 from fgpyo.fasta.builder import FastaBuilder
 from fgpyo.sequence import reverse_complement
 
-from prymer.api import MinOptMax
-from prymer.api import Oligo
-from prymer.api import PrimerPair
-from prymer.api import Span
 from prymer.api import picking
 from prymer.api.melting import calculate_long_seq_tm
+from prymer.model import MinOptMax
+from prymer.model import Oligo
+from prymer.model import PrimerPair
+from prymer.model import Span
 from prymer.primer3 import PrimerAndAmpliconWeights
