+++ tests/api/test_variant_lookup.py
@@ -12,8 +12,6 @@ from fgpyo.vcf.builder import VcfFieldNumber
 from fgpyo.vcf.builder import VcfFieldType
 from pysam import VariantRecord
 
-from prymer.api.span import Span
-from prymer.api.span import Strand
 from prymer.api.variant_lookup import FileBasedVariantLookup
 from prymer.api.variant_lookup import SimpleVariant
 from prymer.api.variant_lookup import VariantOverlapDetector
@@ -21,6 +19,8 @@ from prymer.api.variant_lookup import VariantType
 from prymer.api.variant_lookup import cached
 from prymer.api.variant_lookup import calc_maf_from_filter
 from prymer.api.variant_lookup import disk_based
+from prymer.model import Span
+from prymer.model import Strand
 
 
 @pytest.mark.parametrize(
