+++ tests/primer3/test_primer3_parameters.py
@@ -2,7 +2,7 @@ from dataclasses import replace
 
 import pytest
 
-from prymer.api.minoptmax import MinOptMax
+from prymer.model import MinOptMax
 from prymer.primer3.primer3_input_tag import Primer3InputTag
 from prymer.primer3.primer3_parameters import PrimerAndAmpliconParameters
 from prymer.primer3.primer3_parameters import ProbeParameters
